@@ -1,24 +1,26 @@
 source 'https://rubygems.org'
 
+ruby File.read('.ruby-version').strip
+
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 4.1.8'
+gem 'rails', '~> 5.0.0'
 
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 4.0.3'
+gem 'sass-rails'
 # Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
+gem 'uglifier'
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 # gem 'therubyracer',  platforms: :ruby
 
 # Use jquery as the JavaScript library
 gem 'jquery-rails'
 
-gem 'bcrypt', '~> 3.1.7'
+gem 'bcrypt'
 
 gem 'geminabox', '>= 0.13.5'
-gem 'omniauth', '~> 1.1'
-gem 'omniauth-github', '~> 1.1'
-gem 'omniauth-google-oauth2', '~> 0.2'
+gem 'omniauth-github'
+gem 'omniauth-google-oauth2'
+gem 'sinatra', '2.0.0.beta2'
 gem 'bootstrap-sass'
 gem 'font-awesome-sass'
 gem 'unicorn-rails'
