@@ -1,4 +1,4 @@
-class AddUploader < ActiveRecord::Migration
+class AddUploader < ActiveRecord::Migration[4.2]
   def change
     create_table :uploaders do |t|
       t.string :gem_name, :gem_version, null: false
