@@ -55,13 +55,13 @@ class DevicesControllerTest < ActionController::TestCase
 
     when_not_logged_in do
       it 'denies access' do
-        post :create, params
+        post :create, params: params
         assert_access_denied
       end
     end
 
     when_logged_in_as(:non_admin, :admin) do
-      before { post :create, params }
+      before { post :create, params: params }
 
       describe 'when the device can not be saved' do
         let(:params) do
@@ -94,7 +94,7 @@ class DevicesControllerTest < ActionController::TestCase
   describe 'destroy' do
     when_not_logged_in do
       it 'denies access' do
-        delete :destroy, :id => 1
+        delete :destroy, params: {:id => 1}
         assert_access_denied
       end
     end
@@ -103,7 +103,7 @@ class DevicesControllerTest < ActionController::TestCase
       let(:device) { current_user.devices.last }
 
       it 'destroys the device' do
-        delete :destroy, :id => device.id
+        delete :destroy, params: {:id => device.id}
         Device.where(:id => device.id).must_be :empty?
       end
     end
